// src/requests.d.ts
export declare const requests: {
    data: {
        createEntity: string;
        project: string;
        getEntities: string;
        getEntityById: string;
        addTagsToEntity: string;
        removeTagsFromEntity: string;
        getTags: string;
        updateTag: string;
        getAppData: string;
        updateAppData: string;
        createFile: string;
        getFileById: string;
        getFiles: string;
    };
    toast: {
        show: string;
    };
    menu: {
        populate: string;
    };
    events: {
        menuItemClicked: string;
        menuClicked: string;
    };
};